image and album model complete, both saved to mongo
