+++ image-gallery/server/lib/models/album.js
@@ -2,10 +2,13 @@ const mongoose = require ('mongoose');
 const Schema = mongoose.Schema;
 
 const albumSchema = new Schema({
-    name: String,
-    images:[{
-        title: String,
-        description: String,
+    name:{
+        type:String,
+        required: true,
+    } ,
+    images: [{
+        title:String,
+        description:String,
         url: String,
         wiki: String
     }]
